refactor: extract admin user management handlers to layered architecture
- Create api/admin/users.rs
- Extend repositories/user_repository.rs with 11 admin methods
- Move UserWithBillingInfo struct from db/queries.rs to user_repository.rs
- Remove 10 SQL functions from db/queries.rs
- Remove 6 handlers from router.rs
- Promote sync_link_mapping_from_link to pub(crate) in router.rs
