+++ src/api/admin/mod.rs
@@ -1,2 +1,3 @@
 pub mod api_keys;
 pub mod blacklist;
+pub mod users;
