+++ src/router.rs
@@ -54,7 +54,7 @@ fn hash_ip(ip: &str) -> String {
     format!("{:x}", hasher.finish())
 }
 
-async fn resolved_forward_for_link(db: &D1Database, link: &Link) -> bool {
+pub(crate) async fn resolved_forward_for_link(db: &D1Database, link: &Link) -> bool {
     if let Some(value) = link.forward_query_params {
         value
     } else {
@@ -64,7 +64,7 @@ async fn resolved_forward_for_link(db: &D1Database, link: &Link) -> bool {
     }
 }
 
-async fn sync_link_mapping_from_link(
+pub(crate) async fn sync_link_mapping_from_link(
     db: &D1Database,
     kv_store: &worker::kv::KvStore,
     link: &Link,
@@ -1985,105 +1985,7 @@ pub async fn handle_oauth_callback(req: Request, ctx: RouteContext<()>) -> Resul
     Ok(Response::empty()?.with_status(302).with_headers(headers))
 }
 
-#[utoipa::path(
-    get,
-    path = "/api/admin/users",
-    tag = "Admin",
-    summary = "List all users",
-    params(
-        ("page" = Option<i64>, Query, description = "Page number"),
-        ("limit" = Option<i64>, Query, description = "Items per page (max 100)"),
-    ),
-    responses(
-        (status = 200, description = "Paginated list of users"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_list_users(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    // Parse pagination params
-    let url = req.url()?;
-    let page: i64 = url
-        .query()
-        .and_then(|q| {
-            q.split('&')
-                .find(|s| s.starts_with("page="))
-                .and_then(|s| s.split('=').nth(1))
-                .and_then(|s| s.parse().ok())
-        })
-        .unwrap_or(1);
-
-    let limit: i64 = url
-        .query()
-        .and_then(|q| {
-            q.split('&')
-                .find(|s| s.starts_with("limit="))
-                .and_then(|s| s.split('=').nth(1))
-                .and_then(|s| s.parse().ok())
-        })
-        .unwrap_or(50)
-        .min(100); // Cap at 100 items per page to prevent DoS via unbounded queries
-
-    let offset = (page - 1) * limit;
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-    let users = db::get_all_users_with_billing_info(&db, limit, offset).await?;
-    let total = db::get_user_count(&db).await?;
-
-    Response::from_json(&serde_json::json!({
-        "users": users,
-        "total": total,
-        "page": page,
-        "limit": limit,
-    }))
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/admin/users/{id}",
-    tag = "Admin",
-    summary = "Get a user",
-    params(("id" = String, Path, description = "User ID")),
-    responses(
-        (status = 200, description = "User details"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-        (status = 404, description = "User not found"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_get_user(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    let user_id = ctx
-        .param("id")
-        .ok_or_else(|| Error::RustError("Missing user ID".to_string()))?;
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-    let user = db::get_user_by_id(&db, user_id).await?;
-
-    match user {
-        Some(user) => Response::from_json(&user),
-        None => Response::error("User not found", 404),
-    }
-}
+// handle_admin_list_users and handle_admin_get_user moved to api/admin/users.rs
 
 #[utoipa::path(
     post,
@@ -2508,357 +2410,8 @@ pub async fn handle_admin_sync_link_kv(req: Request, ctx: RouteContext<()>) -> R
 
 // Blacklist handlers moved to api/admin/blacklist.rs
 
-#[utoipa::path(
-    put,
-    path = "/api/admin/users/{id}/suspend",
-    tag = "Admin",
-    summary = "Suspend a user",
-    params(("id" = String, Path, description = "User ID")),
-    responses(
-        (status = 200, description = "User suspended"),
-        (status = 400, description = "Cannot suspend self or last admin"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-        (status = 404, description = "User not found"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_suspend_user(
-    mut req: Request,
-    ctx: RouteContext<()>,
-) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    let target_user_id = ctx
-        .param("id")
-        .ok_or_else(|| Error::RustError("Missing user ID".to_string()))?
-        .to_string();
-
-    // Parse request body
-    let body: serde_json::Value = match req.json().await {
-        Ok(body) => body,
-        Err(e) => return Response::error(format!("Invalid JSON: {}", e), 400),
-    };
-
-    let reason = match body.get("reason").and_then(|r| r.as_str()) {
-        Some(r) => r.to_string(),
-        None => return Response::error("Missing 'reason' field", 400),
-    };
-
-    // Safety guard: Cannot suspend self
-    if target_user_id == user_ctx.user_id {
-        return Response::from_json(&serde_json::json!({
-            "success": false,
-            "message": "Cannot suspend yourself"
-        }));
-    }
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    // Safety guard: Cannot suspend the last admin
-    let admin_count = db::get_admin_count(&db).await?;
-    if admin_count <= 1 {
-        // Check if target user is admin
-        if let Some(target_user) = db::get_user_by_id(&db, &target_user_id).await?
-            && target_user.role == "admin"
-        {
-            return Response::error("Cannot suspend the last admin", 400);
-        }
-    }
-
-    let target_user = match db::get_user_by_id(&db, &target_user_id).await? {
-        Some(user) => user,
-        None => return Response::error("User not found", 404),
-    };
-    let org_links = db::get_links_by_org(&db, &target_user.org_id).await?;
-
-    // Suspend user
-    db::suspend_user(&db, &target_user_id, &reason, &user_ctx.user_id).await?;
-
-    // Disable all links for the user's organization
-    let disabled_count = db::disable_all_links_for_org(&db, &target_user.org_id).await?;
-
-    let kv = ctx.kv("URL_MAPPINGS")?;
-    for link in org_links {
-        if matches!(link.status, LinkStatus::Blocked) {
-            continue;
-        }
-        let mut disabled_link = link;
-        disabled_link.status = LinkStatus::Disabled;
-        sync_link_mapping_from_link(&db, &kv, &disabled_link).await?;
-    }
-
-    // Invalidate all sessions for the user
-    // Note: In a production system, we'd need to track user sessions and delete them
-    // For now, we'll rely on the suspended_at check in auth middleware
-
-    Response::from_json(&serde_json::json!({
-        "success": true,
-        "message": "User suspended successfully",
-        "disabled_links": disabled_count
-    }))
-}
-
-#[utoipa::path(
-    put,
-    path = "/api/admin/users/{id}/unsuspend",
-    tag = "Admin",
-    summary = "Unsuspend a user",
-    params(("id" = String, Path, description = "User ID")),
-    responses(
-        (status = 200, description = "User unsuspended"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_unsuspend_user(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    let target_user_id = ctx
-        .param("id")
-        .ok_or_else(|| Error::RustError("Missing user ID".to_string()))?
-        .to_string();
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-    db::unsuspend_user(&db, &target_user_id).await?;
-
-    Response::from_json(&serde_json::json!({
-        "success": true,
-        "message": "User unsuspended successfully"
-    }))
-}
-
-#[utoipa::path(
-    delete,
-    path = "/api/admin/users/{id}",
-    tag = "Admin",
-    summary = "Delete a user",
-    params(("id" = String, Path, description = "User ID")),
-    responses(
-        (status = 200, description = "User deleted"),
-        (status = 400, description = "Cannot delete self or last org admin"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-        (status = 404, description = "User not found"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_delete_user(mut req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    let target_user_id = ctx
-        .param("id")
-        .ok_or_else(|| Error::RustError("Missing user ID".to_string()))?
-        .to_string();
-
-    // Safety guard: Cannot delete self
-    if target_user_id == user_ctx.user_id {
-        return Response::error("Cannot delete your own account", 400);
-    }
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    // Get target user info for safety checks
-    let target_user = match db::get_user_by_id(&db, &target_user_id).await? {
-        Some(user) => user,
-        None => return Response::error("User not found", 404),
-    };
-
-    // Safety guard: Cannot delete the last admin in an organization
-    if target_user.role == "admin"
-        && db::is_last_admin_in_org(&db, &target_user_id, &target_user.org_id).await?
-    {
-        return Response::error("Cannot delete the last admin in an organization", 400);
-    }
-
-    // Parse request body for confirmation
-    let body: serde_json::Value = match req.json().await {
-        Ok(body) => body,
-        Err(e) => return Response::error(format!("Invalid JSON: {}", e), 400),
-    };
-
-    let confirmation = body.get("confirmation").and_then(|c| c.as_str());
-    if confirmation != Some("DELETE") {
-        return Response::error("Must provide 'confirmation': 'DELETE' in request body", 400);
-    }
-
-    let user_links = db::get_links_by_creator(&db, &target_user_id).await?;
-    let kv = ctx.kv("URL_MAPPINGS")?;
-    for link in &user_links {
-        kv::delete_link_mapping(&kv, &link.org_id, &link.short_code).await?;
-    }
-
-    // Delete user and all their data
-    let (user_count, links_count, analytics_count) =
-        db::delete_user(&db, &target_user_id, &user_ctx.user_id).await?;
-
-    if user_count == 0 {
-        return Response::error("Failed to delete user", 500);
-    }
-
-    Response::from_json(&serde_json::json!({
-        "success": true,
-        "message": "User and all associated data deleted successfully",
-        "deleted_user_count": user_count,
-        "deleted_links_count": links_count,
-        "deleted_analytics_count": analytics_count
-    }))
-}
-
-#[utoipa::path(
-    put,
-    path = "/api/admin/users/{id}",
-    tag = "Admin",
-    summary = "Update user role",
-    params(("id" = String, Path, description = "User ID")),
-    responses(
-        (status = 200, description = "Updated user"),
-        (status = 400, description = "Invalid role or cannot demote last admin"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-        (status = 404, description = "User not found"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_update_user_role(
-    mut req: Request,
-    ctx: RouteContext<()>,
-) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    let target_user_id = ctx
-        .param("id")
-        .ok_or_else(|| Error::RustError("Missing user ID".to_string()))?
-        .to_string();
-
-    // Safety guard: Cannot modify own role
-    if target_user_id == user_ctx.user_id {
-        return Response::error("Cannot modify your own role", 400);
-    }
-
-    // Parse request body
-    let body: serde_json::Value = match req.json().await {
-        Ok(body) => body,
-        Err(e) => return Response::error(format!("Invalid JSON: {}", e), 400),
-    };
-
-    let new_role = match body.get("role").and_then(|v| v.as_str()) {
-        Some(role) => {
-            if role != "admin" && role != "member" {
-                return Response::error("Role must be 'admin' or 'member'", 400);
-            }
-            role.to_string()
-        }
-        None => return Response::error("Missing 'role' field", 400),
-    };
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    // Get target user info for safety checks
-    let target_user = match db::get_user_by_id(&db, &target_user_id).await? {
-        Some(user) => user,
-        None => return Response::error("User not found", 404),
-    };
-
-    // Safety guard: Cannot demote the last admin
-    if new_role == "member" && target_user.role == "admin" {
-        let admin_count = match db::get_admin_count(&db).await {
-            Ok(count) => count,
-            Err(e) => {
-                console_log!("Error checking admin count: {}", e);
-                return Response::error("Failed to verify admin count", 500);
-            }
-        };
-
-        if admin_count <= 1 {
-            return Response::error("Cannot demote the last admin user", 400);
-        }
-    }
-
-    // Update user role
-    match db::update_user_role(&db, &target_user_id, &new_role).await {
-        Ok(_) => {
-            console_log!(
-                "{}",
-                serde_json::json!({
-                    "event": "user_role_updated",
-                    "target_user_id": target_user_id,
-                    "old_role": target_user.role,
-                    "new_role": new_role,
-                    "admin_user_id": user_ctx.user_id,
-                    "level": "info"
-                })
-            );
-
-            // Get updated user data
-            let updated_user = match db::get_user_by_id(&db, &target_user_id).await {
-                Ok(Some(user)) => Ok(user),
-                Ok(None) => Err(Error::RustError(
-                    "Failed to retrieve updated user".to_string(),
-                )),
-                Err(e) => Err(e),
-            }?;
-
-            Ok(Response::from_json(&serde_json::json!({
-                "id": updated_user.id,
-                "email": updated_user.email,
-                "name": updated_user.name,
-                "avatar_url": updated_user.avatar_url,
-                "oauth_provider": updated_user.oauth_provider,
-                "oauth_id": updated_user.oauth_id,
-                "org_id": updated_user.org_id,
-                "role": updated_user.role,
-                "created_at": updated_user.created_at,
-                "suspended_at": updated_user.suspended_at,
-                "suspension_reason": updated_user.suspension_reason,
-                "suspended_by": updated_user.suspended_by,
-            }))?)
-        }
-        Err(e) => {
-            console_log!(
-                "{}",
-                serde_json::json!({
-                    "event": "user_role_update_failed",
-                    "target_user_id": target_user_id,
-                    "new_role": new_role,
-                    "error": e.to_string(),
-                    "admin_user_id": user_ctx.user_id,
-                    "level": "error"
-                })
-            );
-            Response::error("Failed to update user role", 500)
-        }
-    }
-}
+// handle_admin_suspend_user, handle_admin_unsuspend_user, handle_admin_delete_user,
+// handle_admin_update_user_role moved to api/admin/users.rs
 
 #[utoipa::path(
     get,
